Uzum_Market_With_Spring_Bood

+++ src/main/java/com/example/demo/entity/PromoCode.java
@@ -44,7 +44,7 @@ public class PromoCode {
 
     @NotNull
     @Positive
-    private Double discount;
+    private Short discount;
 
     @Builder.Default
     private boolean active=true;
